chore: update to node 14
- update to node 14
- fix circular dependencies related to update
- fix karma timeout
- remove deprecated ngcc calls
