+++ .nvmrc
@@ -1,1 +1,1 @@
-v12.16.0
+v14.16.0
