+++ apps/website/package.json
@@ -34,8 +34,6 @@
     "markdownlint-cli": "^0.24.0",
     "node-sass": "^4.13.1",
     "sass-loader": "^8.0.2",
-    "stylus": "0.54.8",
-    "stylus-loader": "^3.0.2",
     "vuepress": "^1.8.0",
     "vuepress-plugin-clean-urls": "^1.1.2",
     "vuepress-plugin-container": "^2.1.5",
