+++ packages/angular/package.json
@@ -1,7 +1,7 @@
 {
   "name": "angular",
   "private": true,
-  "version": "5.0.3",
+  "version": "0.0.0",
   "scripts": {
     "start": "ng build clr-angular --watch",
     "start:dev": "ng build clr-angular --watch --prod",
@@ -13,15 +13,12 @@
     "serve:dark": "ng serve dev -c dark",
     "storybook": "start-storybook -s .storybook/public -p 6006",
     "storybook:build": "build-storybook -s .storybook/public -c .storybook -o ../../dist/website/storybook/angular --quiet",
-    "postinstall": "ngcc && node ../../scripts/ngcc-clean",
-    "build": "npm-run-all build:lib:cds build:lib:clr build:package:cds build:package:clr build:update:cds build:update:clr",
+    "build": "npm-run-all build:lib:cds build:lib:clr build:update:cds build:update:clr",
     "build:preinstall": "yarn run build:preinstall:cds && yarn run build:preinstall:clr",
-    "build:preinstall:cds": "mkdir -p ../../dist/angular/",
-    "build:preinstall:clr": "mkdir -p ../../dist/clr-angular/",
-    "build:package:clr": "replace '@VERSION' $npm_package_version ./dist/clr-angular/package.json",
-    "build:package:cds": "replace '@VERSION' $npm_package_version ./dist/cds-angular/package.json",
+    "build:preinstall:cds": "mkdir -p ../../dist/angular/ && cp ./projects/cds-angular/package.json ../../dist/angular/",
+    "build:preinstall:clr": "mkdir -p ../../dist/clr-angular/ && cp ./projects/clr-angular/package.json ../../dist/clr-angular/",
     "build:update:clr": "node ../../scripts/update-local-packages.js /packages/angular/dist/clr-angular/package.json",
-    "build:update:cds": "node ../../scripts/update-local-packages.js /packages/angular/dist/cds-angular/package.json",
+    "build:update:cds": "node ../../scripts/update-local-packages.js /packages/angular/dist/angular/package.json",
     "build:lib:clr": "ng build clr-angular --prod",
     "build:lib:cds": "yarn run generate:core-ng-modules && ng build cds-angular --prod",
     "build:adoption": "ng build clarity-adoption --prod",
@@ -53,6 +50,8 @@
     "@angular/platform-browser-dynamic": "11.0.0",
     "@angular/router": "11.0.0",
     "@cds/core": "./dist/core",
+    "@cds/angular": "./dist/angular",
+    "@clr/angular": "./dist/clr-angular",
     "@clr/icons": "./dist/clr-icons",
     "@clr/ui": "./dist/clr-ui",
     "css-vars-ponyfill": "2.3.1",
@@ -104,7 +103,7 @@
     "karma-parallel": "0.3.1",
     "lit-element": "2.3.1",
     "lit-html": "1.2.1",
-    "ng-packagr": "11.0.2",
+    "ng-packagr": "11.2.4",
     "npm-run-all": "4.1.5",
     "postcss-cli": "8.1.0",
     "prettier": "2.0.5",
