+++ packages/angular/projects/clarity-adoption/tsconfig.app.json
@@ -3,7 +3,13 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "../../out-tsc/adoption",
-    "types": []
+    "types": [],
+    "paths": {
+      "@clr/angular": ["dist/clr-angular"],
+      "@clr/angular/*": ["dist/clr-angular/*"],
+      "@cds/angular": ["dist/angular"],
+      "@cds/angular/*": ["dist/angular/*"]
+    }
   },
   "files": ["src/main.ts", "src/polyfills.ts"],
   "exclude": ["**/*.html"],
