+++ packages/angular/projects/clr-angular/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@clr/angular",
-  "version": "@VERSION",
+  "version": "5.0.3",
   "description": "Angular components for Clarity",
   "homepage": "https://clarity.design",
   "keywords": ["ng-add", "clarity", "angular", "components"],
