+++ packages/core/package.json
@@ -86,12 +86,12 @@
     "eslint-config-prettier": "6.11.0",
     "eslint-plugin-jasmine": "4.1.1",
     "fs-extra": "8.1.0",
-    "karma": "4.4.1",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "2.0.5",
+    "karma": "6.2.0",
+    "karma-chrome-launcher": "3.1.0",
+    "karma-coverage-istanbul-reporter": "3.0.3",
     "karma-htmlfile-reporter": "0.3.8",
-    "karma-jasmine": "2.0.1",
-    "karma-jasmine-html-reporter": "1.4.2",
+    "karma-jasmine": "4.0.1",
+    "karma-jasmine-html-reporter": "1.5.4",
     "lit-analyzer": "1.2.0",
     "npm-run-all": "4.1.5",
     "purgecss": "2.3.0",
