+++ packages/core/src/internal/decorators/i18n.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 VMware, Inc. All Rights Reserved.
+ * Copyright (c) 2016-2021 VMware, Inc. All Rights Reserved.
  * This software is released under MIT license.
  * The full license information can be found in LICENSE in the root directory of this project.
  */
@@ -21,7 +21,7 @@ function standardI18n(descriptor: PropertyDescriptor, element: { key: string }) 
     descriptor,
   };
 
-  return property({ type: Object })(desc, name);
+  return property({ type: Object })(desc);
 }
 
 /**
