+++ packages/core/src/internal/utils/string.ts
@@ -121,50 +121,3 @@ export function removePrefix(str: string, prefix: string): string {
 export function removeSuffix(str: string, suffix: string): string {
   return removePrefixOrSuffix(str, suffix, 'suffix');
 }
-
-type ObjectPropertyNameAndValueTuples =
-  | [string, string]
-  | [string, true]
-  | [string, false]
-  | [string, null]
-  | [string, undefined]
-  | [string, number];
-
-// this utility is a little restrictive on its inputs. it expects a specific format.
-// going outside of that format (like passing an object or function as the value) can yield unintended results.
-// this does NOT eval anything. that would be bad.
-// 'isValid:true status:success': string =>
-// [['isValid', true], ['status', 'success']]: [string, string | number | boolean][]
-export function convertStringPropValuePairsToTuple(propValString: string): ObjectPropertyNameAndValueTuples[] {
-  // starts as a string like... "isValid:true status:success"
-  return propValString
-    .split(' ')
-    .map(str => str.split(':'))
-    .map(pv => {
-      const [propname, propValAsString] = pv;
-
-      if (propValAsString === 'true') {
-        return [propname, true];
-      }
-
-      if (propValAsString === 'false') {
-        return [propname, false];
-      }
-
-      if (propValAsString === 'null') {
-        return [propname, null];
-      }
-
-      if (propValAsString === 'undefined') {
-        return [propname, undefined];
-      }
-
-      if (isNumericString(propValAsString)) {
-        return [propname, +propValAsString];
-      }
-
-      // else it's a string and that's ok
-      return [propname, propValAsString];
-    });
-  // returns as [['isValide', true], ['status', 'success']]
-}
