+++ packages/icons/npm.json
@@ -1,6 +1,6 @@
 {
   "name": "@clr/icons",
-  "version": "@VERSION",
+  "version": "5.0.3",
   "description": "Custom Element Icons for Clarity",
   "homepage": "https://clarity.design",
   "main": "index.js",
