+++ packages/ui/npm.json
@@ -1,6 +1,6 @@
 {
   "name": "@clr/ui",
-  "version": "@VERSION",
+  "version": "5.0.3",
   "description": "CSS for Clarity",
   "homepage": "https://clarity.design",
   "repository": {
