+++ packages/ui/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@clr/ui",
   "private": true,
-  "version": "5.0.3",
+  "version": "0.0.0",
   "scripts": {
     "clean": "del ./dist",
     "build": "npm-run-all clean build:css build:prefix build:optimize build:package build:src build:update",
@@ -12,7 +12,7 @@
     "build:css": "sass --precision=8 src/main.scss ./dist/clr-ui/clr-ui.css; sass --precision=8 src/dark-theme.scss ./dist/clr-ui/clr-ui-dark.css",
     "build:prefix": "postcss ./dist/clr-ui/clr-ui.css -o ./dist/clr-ui/clr-ui.css; postcss ./dist/clr-ui/clr-ui-dark.css -o ./dist/clr-ui/clr-ui-dark.css",
     "build:optimize": "csso -i ./dist/clr-ui/clr-ui.css -o ./dist/clr-ui/clr-ui.min.css -s file --no-restructure; csso -i ./dist/clr-ui/clr-ui-dark.css -o ./dist/clr-ui/clr-ui-dark.min.css -s file --no-restructure",
-    "build:package": "cpy npm.json README.md ./dist/clr-ui/; mv ./dist/clr-ui/npm.json ./dist/clr-ui/package.json; replace '@VERSION' $npm_package_version ./dist/clr-ui/package.json; "
+    "build:package": "cpy npm.json README.md ./dist/clr-ui/; mv ./dist/clr-ui/npm.json ./dist/clr-ui/package.json;"
   },
   "repository": {
     "type": "git",
